tools: setup an instance PostgreSQL database
The postgresql.py script starts a new instance of PostgreSQL
that is deleted after termination. The script also prints
the current PostgreSQL's connection parameters and pid of
the running script for controlled terminating.

The PostgreSQL temporary database is created in the current
directory as tmp-postgresql and it is removed on exit.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>