gateways: introduce timezone
Each gateway has a time zone defined. It is the name that is recognizable
by the application server. The currently used time zone implementation is
based on the ICU library. The used time zones are mostly based on a pair
continent name and city name (e.g. Europe/Prague).

The default time zone is however just 'GMT'.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>