api: fix queries to use now_utc() instead of NOW()
The pure NOW() returns timestamp with timezone and when converted
to timestamp, its value stays the some, just its zone is cleared
to 'UTC'. Thus, we have +1 shift of time which is wrong.

E.g. we update a device, storing at 19:55 +01:00. The time is converted
to timestamp as 19:55 +00:00. We have wrong time in the database.
When reading it and interpreting in the +01:00 zone, we get 20:55 +01:00.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>