Constraint sensor data to avoid NaN
The tables:

* beeeon.sensor_history_recent
* beeeon.sensor_history_last
* beeeon.control_fsm
* beeeon.control_fsm_last

now refuses to store NaN values. Replace them by NULL. The NaN was used to denote that
the target sensor is "broken", "disconnected", "misconfigured", etc. This make sense
for sensor_history but it does not make any sense for controls_fsm.

Prior application of this change, it might be needed to fix the database by
query

  UPDATE beeeon.sensor_history_recent
  SET
      value = NULL
  WHERE
      value = 'NaN'::real;

  UPDATE beeeon.sensor_history_last
  SET
      value = NULL
  WHERE
      value = 'NaN'::real;

Similarly for controls_fsm and control_fsm_last but this is not very probable to be
an issue.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>