beeeon.sensor_history_raw: drop unneeded columns
This is the most crucial step in the commit series. We drop the
columns gateway_id, device_id and module_id from the sensor_history_raw
table. The table now references a sensor via the refid surrogate key.

This change might safe abour 30 % of storage data and what is important,
it makes the primary index about 30 % smaller. This can accelerate queries
a bit and opens ways to other optimizations. E.g. it is not much more
desirable to create partial indexes.

The reverting process is a bit complicated as it reverts to the original
physical layout of the table sensor_history_raw (ALTER...ADD COLUMN
appends columns and cannot insert them on their original position).
To preserve the original database state, we drop the value and at columns
and recreate them (safely). This process requires to recreate certain
functions (they must be however changed anyway), view sensor_history
and its rule and the sensor_history_raw check_value_not_nan constraint.

Anyway, after deploying or reverting this change, it is desirable to
call CLUSTER (or anything similar) on the sensor_history_raw table using
its PK to optimize the table organization (quite an expensive operation).

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>