api: optimize queries sensors_history.huge_fetch_*
After the sensor_history_raw is changed to use refid only, we can optimize
queries to take advantage of this fact. Using the view sensor_history is
inefficient anyway.

We are using the sensor_history_raw directly and make a very cheap query
to the sensors table to translate gateway_id, device_id and module_id
to refid.

The query huge_fetch_raw is expected to be accelerated a lot as it depends
mostly on the number of I/O operations that should now be significantly
lower.

The query huge_fetch_agg should be faster now and its bottleneck is sorting
of the resulted data by the aggregated timestamp which seems to not be
possible to accelerate via any index.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>