api: optimize sensors_history.huge_fetch_agg
Get rid of the CTE that is really unneeded in this query. Its original
purpose was to avoid ODBC to fail because we ware referencing the $6
parameter multiple times. This is no longer true and we can avoid the
CTE (probably unneeded I/O).

The type of $6 is also changed to integer as the bigint is simply too
big for this purpose.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>