api: new query sensors_history.huge_fetch_enum_agg
The query aggregates enum values into a histogram-like format.
For each time (at) aligned to the given interval, a sequence
of raws with value and its frequency is computed.

The result set contains a row for each distinct (at, value) pair.
The row is actually a triple (at, value, freq).

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>