api: divide query update_request to 2 specific ones
The query update_request might unintentionally lead to an unwanted
write into database where a the finished_at is overwritten even if
it is already set. To avoid that situation, the original query is
divided into 2 queries that do exactly one thing and only if it is
an appropriate reaction. I.e. already finished request is not updated.
Already accepted request is not accepted again.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>