api: new query gateways_messages.insert
Insert a new message into the gateways_messages table and preserve
maximum of N messages per gateway. The N is configurable per-query
and thus it is not a hard constant of the database but it is given
by the application layer. In this way, it is possible e.g. to have
a different limit of messages per gateway.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>